use byterange to get content length
Jira: VOD-2873

this derives our expected body size by the header Range request
value. now we can fail if the body s3 transfer fails and return
an HTTP error instead of failing silently. we set the header
length so that if it doesn't match then the client (nginx) will
know it failed. also we now get more logging detail on size of
the requested body too.