Setup CI pipeline with Travis
For demo purpose, removed all private repo references to make it a 100%
opensource project. Also cleaned up the old implementation to put
together a clean bare bone structure to rewrite the app.

Also integrated with a couple of code quality and static code analysis
tool to showcase the possibility. Travis CI is set up to work with Ginkgo
(with Gomega) and spits out code coverage details

note: This will not go to `master`. If merged, _this will break the build
and other dependencies_.